Improved debris belt + various minor changes
The debris belt now spawns in a circle, rather than at just two points. This means that the actual belt spawns faster and maintains its shape better. The sun expansion has also been improved; now a red filter is applied to the screen the longer the ending goes on, and eventually the player is just killed. This is to prevent players attempting to outrun the expansion.